Define ObjSpaceViewDir in Fragment Shader
When calling `ObjSpaceViewDir()` Per Vertex from the Vertex Shader there is visible Warbling to Texture3D due to inaccuracies missed
which are correct when calling the same method Per Pixel in the Fragment Shader. Either allow calling of this method from both Shader Programs
or provide a separate implementation for each.